+++ examples/cpp/dds/Benchmark/Benchmark_bigPubSubTypes.cxx
@@ -35,7 +35,7 @@ BenchMarkBigPubSubType::BenchMarkBigPubSubType()
 {
     setName("BenchMarkBig");
     uint32_t type_size = BenchMarkBig_max_cdr_typesize;
-    type_size += eprosima::fastcdr::Cdr::alignment(type_size, 4); /* possible submessage alignment */
+    type_size += static_cast<uint32_t>(eprosima::fastcdr::Cdr::alignment(type_size, 4)); /* possible submessage alignment */
     m_typeSize = type_size + 4; /*encapsulation*/
     m_isGetKeyDefined = false;
     uint32_t keyLength = BenchMarkBig_max_key_cdr_typesize > 16 ? BenchMarkBig_max_key_cdr_typesize : 16;
